CameraHAL : Modify logic of Resolution check for Potrait Resolutions
With the current implementation, the Capabilities were misleading.
This patch uses the existing capabilities but internally swaps width
and height for checking the validity of resolutions.
Added 1920 * 1088 as a supported resolution since this resolution will
be used as 1088 * 1920 for potrait 1080p record scenario.

Signed-off-by: Sathya Prakash M R <a0393677@ti.com>
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>

Patch Set 3:  Simple cherry-pick on current tree.  No changes made.
Signed-off-by: Jean Johnson <a0271255@ti.com>
Change-Id: Ibcca18089c0c46f215a837b43a1fb65f644e0e96