CameraHal: Prevent deadlock in AppCallbackNotifier::stop()
There is a small chance that a stopPreview call can come from
CameraService right around the same time Encoder thread will
send a data callback for the video snapshot. Currently, we
are waiting for the encoder thread to join in AppCallbackNofier::
stop(), so we could deadlock if CameraService locks in
lockIfMessageWanted for the video snapshot.

Instead of waiting for Encoder thread to join, we can make
cancel() block until Encoder thread is done canceling the
encode. After cancel() returns, we can free up the cookies
that we passed to it, so Encoder thread does not need to call
the callback function to AppCallbackNotifier.

Change-Id: I93e21aefc968be810028fa672928fe9e18dff7e4