CameraHAL: RAW + JPEG support
This patch adds functionality for saving both JPEG and respective RAW buffer.
This is done by enabling the video port( to receive RAW ),
and sent data to camera_test via RawCallback.

NOTE: The raw size is defined by mRawWidth and mRawHight.
They set the RAW resolution, which must be the native resolution of the sensor.

Signed-off-by: Georgi Georgiev <x0154504@ti.com>

Rebase.
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>
Signed-off-by: Jean Johnson <a0271255@ti.com>
Change-Id: Ie8cb62e86829fafc6cd3dc303830b40dbeb39cb6