Security: tee_client_api: Update to 1.06e1
The TEE Client API defines a communications API for connecting
Client Applications running in a rich operating environment with
the Secure Services running inside the Trusted Foundations execution
environment.

The TEE Client API is a static library that must be linked with the
calling application or library.

The TEE Client API is used to implement higher-level APIs, such as
cryptography or secure storage.

Change-Id: Ic3c576779a19f25a8585e647b4881905164ad2d7
Signed-off-by: Bryan Buckley <bryan.buckley@ti.com>
Signed-off-by: Yong Zhi <y-zhi@ti.com>
Signed-off-by: Jorge E. Solano <x0062302@ti.com>