hwc: changes for blit framebuffers
Instead of trying to use the SGX swap chain, we're going to use
seperate blit buffers.

The problem with the swapchain is that it is non-trivial to interact
with when attempting to offload SGX compositions to 2D. SGX is still
used in transitions to still needs to interact with the swap chain.

Clear of regions doesn't appear to work yet

Change-Id: I41e4883a54f3921b71981716ee7bce65656c3def
Signed-off-by: Tony Lofthouse <a0741364@ti.com>