hwc: Enable regionizer to work with buffer handle indexes
This patch makes the HWC regionizer code to work with buffer
indexes in the same way as RGZ_OUT_BVCMD_PAINT, the indexes are
stored in the Bltsville buffer descriptions in the virtaddr, the
indexes are determined in the rgz_in operation.

Note: This code doesn't break RGZ_OUT_BVCMD_PAINT with blitmode 1
and 2. Batching is disabled for now since it's not working.

Change-Id: I2e9182d3787b2d7871773f612ae9b5faf68b1e47
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>