hwc: Replacing FBIO_WAITFORVSYNC with sleep
Presently hwc uses FBIO_WAITFORVSYNC to block the
hwc:set() call and also to limit the rate of set call.
But FBIO_WAITFORVSYNC is only valid for Video mode panels.
Hence replacing the FBIO_WAITFORVSYNC with sleep of 16ms
in hwc to limit the rate of hwc:set() to 60fps.

Change-Id: I6db1ea3ecc1dcc82ab736a9e697565b52accfc66
Signed-off-by: Muralidhar Dixit <murali.dixit@ti.com>