hwc: Enable dirty region handling with the regionizer
The regionizer now generates blits only for the regions in the
screen that changes. That is, every subregion in a horizontal
region that contains a dirty buffer handle generates blits,
otherwise that specific subregion is skipped without generating
any blits.

This change handles the scenario where the regionizer receives
multiple times the same geometry and buffer handles and decides
to generate no blits but still there are blit buffers to
synchronize with.

Change-Id: Ia2eab89b6ac25fb39aa66d87c4a2da899cc693a3
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>