hwc: Enable regionizer to work with buffer handle indexes
This patch makes the HWC regionizer code to work with buffer
indexes in the same way as RGZ_OUT_BVCMD_PAINT, the indexes are
stored in the Bltsville buffer descriptions in the virtaddr, the
indexes are determined in the rgz_in operation.

Note: This code doesn't break RGZ_OUT_BVCMD_PAINT with blitmode 1
and 2. Batching is disabled for now since it's not working.

Change-Id: I1b58eb31a6828f9a8d3952b3c96d5074243787b6
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>