CameraHal: Add CameraBuffer (major refactoring)
Add a CameraBuffer type, which contains information corresponding to
an internal CameraHal buffer - how it was allocated, relevant handles,
and any internal data to be associated with the buffer, such as metadata,
OMX indexes, current state.  The simplest possible refactoring was done
to convert the void * used for exchanging buffer information into the
'opaque' field in the CameraBuffer structure.  A few hacks were added
to minimize the size of the refactoring, these will be cleaned up in
upcoming commits.

Adding a private structure for buffers eliminates the need for a lot
of KeyedVectors that are currently in use.  Most of these were left in
place, and will be refactored into the CameraBuffer structure later.
The pAppPrivate field of OMX_BUFFERHEADERTYPE was used to eliminate
another KeyedVector.

Change-Id: I147ad971e4cdc03b2c445b5fba81d0ba83bc5fb3
Signed-off-by: David Schleef <ds@ti.com>
Signed-off-by: Vladimir Petrov <vppetrov@mm-sol.com>
Signed-off-by: Tyler Luu <tluu@ti.com>