camera_test: implement streaming capture
To test streaming capture, use 'uuuu1Sp' to start streaming
capture and then 'S' to toggle.

Changes in this patch:
1. Needed to reduce latency when handling buffer, so had to
remove memcpy and file write in handleBuffer. Instead of memcpy,
we are just saving the graphic buffer and locking it later during
reprocess to get the mapped buffer. File write is disabled during
streaming capture only.
2. Change Defer worker thread to handle requests in a queue instead
of spawning a new thread for each request.

Change-Id: I699aede8a04b2513f92e0d9ed926dd8189ccdcca
Signed-off-by: Tyler Luu <tluu@ti.com>
Signed-off-by: Vladimir Petrov <vppetrov@mm-sol.com>