OMAP: Make a common make file and create omap5.mk
Make a common make file for all common components.
Then create an omap5.mk file for omap5 specific includes.

This will allow the common components to be included an then
platform or product differentiation to occur for the specific OMAP.

Currently the only difference when this patch was made was the module name

Change-Id: Id4d08debaa3003b2471a084618d848a26dee3881
Signed-off-by: Dan Murphy <dmurphy@ti.com>