hwc: enable partial offload of SGX to gc320 if hardware is available
Whilst DSS remains the primary mechanism to perform UI composition
this patch enables offload of some usage of SGX for composition given
the following conditions

- DSS runs out of pipelines
- Layer Pixel formats: RGBA / RGBX / BGRA / BGRX / RGB565
- Layer transformation (Scale / Rot): none

SGX will be used for any other conditions which are not handled by DSS
and GC320.

There is no guarantee this will perform better than SGX at this moment
in time.

This patch also depends on a kernel patch to gcioctl to fail the device
open on hardware without GC320.

Change-Id: Idbc3ee2e2adb09be5ea08fb96c6dfb0a19d0700f
Signed-off-by: Tony Lofthouse <a0741364@ti.com>