hwc: use triple buffer for overlay layers
This is a cheat to discriminate between blit layers and real overlay engine
layers.

In principle overlay layers should be triple buffered as the benefit of
zero copy composition can be a little offset by a buffer staying in the
display pipeline and stalling the renderer for want of a buffer.

However the requisite functionality for HWC to inform SurfaceFlinger to
dynamically triple buffer is not present, this is likely to arrive in
JellyBean.

This patch is benign and dumpsys hints for layers will tell you what is an real
overlay or not.

For an OVERLAY type layer the hints field should have bit 1 set when using dss
overlays, and clear when it was blitted.

Change-Id: Ieede0dc4ea373e903a91c5cd6035a04c790d326d
Signed-off-by: Tony Lofthouse <a0741364@ti.com>