hwc: Added support for ext display back buffer allocation
When the ext display transform is different from the FB,
the mirroring/cloning is achived with the help of back-buffers,
which are allocated in TILER2D space to get the required
transformation. This patch adds support in HWC to detect
the ext display transform and allocate back buffers from
TILER2D space and program dsscomp accordingly.

Change-Id: Ic9d3bc5d7777bad12f96a373c4c64e7c9d2ff3f7
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>