CameraHal: Changes necessitated by slice based processing
This patch reverts the revert commit
http://review.omapzoom.org/#/c/22000/

and enables the Slice mode in camera hal as in original patch
http://review.omapzoom.org/#/c/19749/

Couple of changes are present in addition to original patch:

The initial revert was to due to 60fps issue on OMAP5.
The issue was debugged and confirmed to be caused by combination of
existing code for handling state change when sensor overclocking is
enabled in combination with port enable/disable changes of this patch.
Mutex handling has been updated to improve the scenario (another patch
expected from camera team to manage port status)

Another change has also been incorporated in this patch to address the
logging related compilation issue when debug flags are enabled in camera hal.

Change-Id: I5b78d2bdc3f82133be8447176bebde7f4673441b
Signed-off-by: Prasanna Kumar M.R. <prasanna.kumarmr@ti.com>
Signed-off-by: Anu Sundararajan <sanuradha@ti.com>