CameraHAL: Fix for disable port time out.
[Comment]
There was a port disable timeout in switchToExecuting() as
the port is not enabled state. So while switching the preview
from image preview to video preview with 60fps other than 1080p
resolution, >5sec delay is noticed because of this port disable
time out issue.

This issue is noticed after the below change.
http://review.omapzoom.org/#/c/22417/

This patch enables the preview port during switchToLoaded if
the sensor overclock is set. So that portDisable will not
time out in doSwitchToExecuting().

Change-Id: I1501520337f9c41d68761ddce9abd81328fb7443
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>