hwc: regionizer: Clear regions with no layer intersections
The regionizer was clearing the whole screen even if the layers
blitted where covering all of it. This change prevents this
situation by introducing a screen sized background layer as the
backmost layer, this makes the regionizer to generate regions
for the whole screen. Regions which instersect just with the
background layer generates a clear for that region, in the case
where the layers to be composed cover the whole screen no clear
is done. Performance is greatly improved with this optimization.

Change-Id: I0d978e91bac1f26c8aa2f34c20f093136dd67191
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>
Signed-off-by: a0741364 <tony.lofthouse@ti.com>