hwc: regionizer: Add asynchronous blits support
Regionizer adds the BVFLAG_ASYNC to all the blits, except for the
last one which is made synchronous. The last blit acts like a
fence to let the previous asynchronous blits to finish first.

Support in the GC driver must be present for this feature to work
properly.

Rebased from p-ics-mr1:
822ec12 hwc: regionizer: Add asynchronous blits support

Change-Id: Iec32de51e7d0b95d1b780e60ed4848e8daf7977f
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>