hwc: work-around for panda5 by forcing fake vsync
Panda5 has an HDMI output, but the latest h/w composer api expects
a vsync implementation. A trivial work-around is to go to the previous
api version in which circumstance surfaceflinger will generate 'fake'
vsyncs using a thread and timer.

This can be removed when AVsync is working in the kernel for HDMI

Change-Id: I223db3b69d707cd14f5ab67bedb960ecb3b50f87
Signed-off-by: Dan Murphy <dmurphy@ti.com>