hwc: HACK: wuxga panels will generate unscalable layers
During the transition from the recents panel to a selected application
during the transition a layer which requires a scale down from 1920x1200
to 594x361 is generated. This satisfies the normal limits but with buffers
of this size and also blending another 1920x1200 framebuffer. A VID1
underflow occurs.

This work-around avoids scaling when layers are wider than 720p and
are 4 byte RGB.

These conditions have been reproduced in isolation with ADTF.

Change-Id: If8fbc8f639abb89a0d736a3beabd2c12312ac3ac
Signed-off-by: Tony Lofthouse <a0741364@ti.com>